change to bigint for total_active_stake

+++ db/migrate/20221007070911_create_cluster_stat.rb
@@ -1,7 +1,7 @@
 class CreateClusterStat < ActiveRecord::Migration[6.1]
   def change
     create_table :cluster_stats do |t|
-      t.integer :total_active_stake
+      t.bigint :total_active_stake
       t.string :network
 
       t.timestamps
